Rename files (remove module inception)
